Getting things working with PHP 5.4, mostly
Fixed several instances of <?, which is now deprecated -- you are
supposed to always say <?php.

Removed a bunch of permissions at admin login related to the forums that
we aren't using.

Replaced date() with gmdate(), since you aren't allowed to use date()
without a timezone any more.

Changed the DB connection info to match the trivial defaults that I am
using for development.