Commented out a chunk of code
Much of this is peculiar to Atlantia's St. Aidan, which seems to be a
pretty strange award.

The rest is adding artificial AoAs, PoAs, GoAs, based on the rank of the
given award.  This is *terrible* from a DB perspective: it's duplicating
data, and isn't precisely correct -- it's representing ranks as awards.
And it isn't how the East traditionally thinks of this stuff...